+++ src/astrameter/config/config_loader.py
@@ -559,7 +559,7 @@ def create_esphomenative_powermeter(
 ) -> Powermeter:
     return ESPHomeNative(
         address    = config.get(section, "ADDRESS", fallback=""),
-        port       = config.get(section, "PORT", fallback=""),
+        port       = config.get(section, "PORT", fallback="6053"),
         apiKey     = config.get(section, "API_KEY", fallback=""),
         objectId   = config.get(section, "OBJECT_ID", fallback=""),
         clientInfo = config.get(section, "CLIENT_INFO", fallback="AstraMeter")